+++ docs/scripts/app.js
@@ -13,7 +13,8 @@
 	angular
 		.module('21days', [
 			'tmh.dynamicLocale',
-			'720kb.datepicker'
+			'720kb.datepicker',
+			'ngclipboard'
 		]);
 
 })(window.angular);