+++ docs/scripts/services/date.service.js
@@ -55,7 +55,7 @@
 
 				// Convert the date
 				logService.service(data.service, 'original date is: ' + $date);
-				let date = moment($date);
+				let date = moment($date, appConstant.moment.readableFormat);
 				logService.service(data.service, 'moment original date is: ' + methods.readable(date));
 
 				// Add 21 days